Movendo estilos da tabela
